Separate Node interface from NodeFactory
